test(cockpit): extract smee-doorbell test harness into shared helper
Move duplicated scaffolding (FakeServer/startFakeSmee, checkRunFrame,
issueFrame, fakePrSnapshot, setPrev, waitFor, FAKE_RESOLVED) from
smee-source.integration.test.ts and smee-source.checks-casing.test.ts
into a new non-test helper module at __tests__/helpers/smee-harness.ts.

The helper filename does not match vitest's include glob so it is not
collected as a suite, and its location under __tests__/ keeps it out of
coverage per the config's exclude. Each test file retains its own
per-file vi.mock('@generacy-ai/cockpit', ...) call — module mocks are
hoisted per-file and do not survive extraction.

Mutation re-verified: inlining smee-source.ts:375 as
\`this.prev.get(\`\${ev.repo}#pr#\${ev.number}\`)\` still fails exactly
the two read-mixed rows in smee-source.checks-casing.test.ts. All 15
tests in smee-source.integration.test.ts still pass.
